rubah struktur, rename route, nambah 1 widget
