+++ lib/feature/home/page/my_home_page.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class MyHomePage extends StatelessWidget {
+  const MyHomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Expanded(
+              child: Center(
+                child: Text(
+                  'Home',
+                  style: GoogleFonts.poppins(),
+                ),
+              ),
+            ),
+            IconButton(
+              icon: const Icon(Icons.search),
+              onPressed: () {},
+            ),
+          ],
+        ),
+      ),
+      body: Column(
+        children: [
+          Text(
+            'Discover the most modern furniture',
+            style: TextStyle(
+              fontWeight: FontWeight.w400,
+              fontSize: 20,
+              fontFamily: GoogleFonts.poppins().fontFamily,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
